Version 2.1
* Now when users vote (or do any other poll action), they're redirected
to the page where they voted rather than the poll's thread. The original
behavior can be restored with a setting. There are two exceptions to
this:
* ASB: The member profile pages. On these pages, the user is redirected
to the forum index.
* Standard: Users are redirected to the forum index when editing the
poll, even if they commenced the action from the Portal.
* Adde a setting to disable display of the poll if it is closed.
* Standard: Fixed some display issues and made other improvements to the
compact layout.
* Standard: Templates are now added to the global template list when
applicable, so that they aren't generating unnecessary database queries.
* Improved the code's compliance with the MyBB Development Standards.
* General code optimizations.